添加了追踪机器人和A*搜索算法

+++ src/entity/Projectile.java
@@ -6,7 +6,7 @@ import com.again.Sprite;
 
 public class Projectile extends Entity{
 
-	protected final int xOrigin, yOrigin;
+	protected final double xOrigin, yOrigin;
 	protected double angle;
 	protected Sprite sprite;
 	protected double x, y;
@@ -16,7 +16,7 @@ public class Projectile extends Entity{
 	
 	protected final Random random = new Random();
 	
-	public Projectile(int x, int y, double dir)
+	public Projectile(double x, double y, double dir)
 	{
 		xOrigin = x;
 		yOrigin = y;
